niche code refactoring
