+++ core/src/main/java/com/netflix/conductor/core/execution/archival/Endpoint.java
@@ -1,4 +1,4 @@
-package com.netflix.conductor.core.execution.tasks;
+package com.netflix.conductor.core.execution.archival;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
