+++ core/src/main/java/com/netflix/conductor/core/execution/archival/ReminderApi.java
@@ -1,4 +1,4 @@
-package com.netflix.conductor.core.execution.tasks;
+package com.netflix.conductor.core.execution.archival;
 
 import retrofit2.Call;
 import retrofit2.http.Body;
