+++ core/src/main/java/com/netflix/conductor/core/execution/archival/ReminderPayload.java
@@ -1,5 +1,4 @@
-package com.netflix.conductor.core.execution.tasks;
-
+package com.netflix.conductor.core.execution.archival;
 
 import java.io.Serializable;
 import java.util.HashMap;
